+++ src/main/java/com/cts/FoodChainX/repository/NotificationRepository.java
@@ -4,7 +4,7 @@ import com.cts.FoodChainX.model.Notification;
 import org.springframework.data.jpa.repository.JpaRepository;
 import java.util.List;
 
-public interface NotificationRepository extends JpaRepository<Notification, Integer> {
+public interface NotificationRepository extends JpaRepository<Notification, Long> {
 
     // Navigate from Notification -> User -> userId
     List<Notification> findByUserUserIdOrderByCreatedDateDesc(Long userId);
