+++ src/main/java/com/cts/FoodChainX/repository/WarehouseRepository.java
@@ -6,17 +6,17 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.cts.FoodChainX.model.Warehouse;
 
-public interface WarehouseRepository extends JpaRepository<Warehouse, Integer> {
+public interface WarehouseRepository extends JpaRepository<Warehouse, Long> {
 
     
-    List<Warehouse> findByDistributorId(Integer distributorId);
+    List<Warehouse> findByDistributorId(Long distributorId);
 
     
     List<Warehouse> findByStatus(String status);
 
     
-    List<Warehouse> findByDistributorIdAndStatus(Integer distributorId, String status);
+    List<Warehouse> findByDistributorIdAndStatus(Long distributorId, String status);
 
     
-    List<Warehouse> findByDistributorIdOrderByCapacityDesc(Integer distributorId);
+    List<Warehouse> findByDistributorIdOrderByCapacityDesc(Long distributorId);
 }
